+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Fixed invisible bug for spectators that joined while ingame
 * Fixed trail particle are random even on bow shoot  
 * Changed config.yml Item values of states and added all states
+* Added trails blacklist
 
 ### 1.7.5 Release (01.02.2021 - 14.03.2021)
 * Added modifiable arena state names to language